feat: add date normalization utility
